requirements.txt
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.30.6
python-pptx==0.6.23
